+++ lib/src/deepgram_live_listener.dart
@@ -8,7 +8,8 @@ import 'package:web_socket_channel/web_socket_channel.dart';
 /// Class used to transcribe live audio streams.
 class DeepgramLiveListener {
   /// Create a live transcriber with a start and close method
-  DeepgramLiveListener(this.apiKey, {required this.inputAudioStream, this.queryParams});
+  DeepgramLiveListener(this.apiKey,
+      {required this.inputAudioStream, this.queryParams});
 
   /// if transcriber was closed
   bool _isClosed = false;
@@ -29,7 +30,8 @@ class DeepgramLiveListener {
   final Map<String, dynamic>? queryParams;
 
   final String _baseLiveUrl = 'wss://api.deepgram.com/v1/listen';
-  final StreamController<DeepgramListenResult> _outputTranscriptStream = StreamController<DeepgramListenResult>();
+  final StreamController<DeepgramListenResult> _outputTranscriptStream =
+      StreamController<DeepgramListenResult>();
   late WebSocketChannel _wsChannel;
   Timer? _keepAliveTimer;