+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 3.0.0
+**Breaking changes**
+now you can find all method with their respectives subclasses e.g:
+- `deepgram.transcribeFromFile()` -> `deepgram.listen.file()`
+- `deepgram.speakFromText()` -> `deepgram.speak.text()`
+  
+- added support for  `deepgram.speak.live()` - `deepgram.toWav()` makes the audio data readable 
+- updated examples (flutter and dart)
+- added mandatory default query params to every streaming method (encoding, sampleRate)
+
 ## 2.3.2
 - Stop keepalive timer on close (thanks @PcolBP)